Refactor top level package name
