+++ .gitignore
@@ -2,3 +2,4 @@
 .vscode/
 __pycache__/
 /*.csv
+build/
