--- src/visual_pandas_curator.egg-info/SOURCES.txt
@@ -1,16 +1,0 @@
-LICENSE.txt
-NOTICE.txt
-README.md
-pyproject.toml
-src/__init__.py
-src/confirm_selection_window.py
-src/dataframe_plot_widget.py
-src/histogram_plot_widget.py
-src/main_window.py
-src/mpl_widget.py
-src/time_series_data_plot_widget.py
-src/visual_pandas_curator.egg-info/PKG-INFO
-src/visual_pandas_curator.egg-info/SOURCES.txt
-src/visual_pandas_curator.egg-info/dependency_links.txt
-src/visual_pandas_curator.egg-info/requires.txt
-src/visual_pandas_curator.egg-info/top_level.txt
