--- src/visual_pandas_curator.egg-info/requires.txt
@@ -1,8 +1,0 @@
-pandas>=1.1.5
-PyQt5==5.14.0
-matplotlib>=3.7.1
-seaborn>=0.11.1
-overrides>=7.3.1
-
-[format]
-black>=23.3.0
