--- src/visual_pandas_curator.egg-info/top_level.txt
@@ -1,7 +1,0 @@
-__init__
-confirm_selection_window
-dataframe_plot_widget
-histogram_plot_widget
-main_window
-mpl_widget
-time_series_data_plot_widget
