+++ visual_pandas_curator/dataframe_plot_widget.py
@@ -3,9 +3,9 @@ __maintainer__ = "Manuel Yves Galliker"
 __license__ = "Apache-2.0"
 
 try:
-    from src.mpl_widget import MplWidget
+    from visual_pandas_curator.mpl_widget import MplWidget
 except:
-    from visual_dataframe_selector.src.mpl_widget import MplWidget
+    from visual_pandas_curator.src.mpl_widget import MplWidget
 
 from abc import abstractmethod
 import pandas as pd
